+++ src/compiler/link.c
@@ -639,7 +639,7 @@ static bool fread_ty (FILE *f, Ty_ty *ty)
         case Ty_record:
         case Ty_pointer:
         case Ty_string:
-        case Ty_void:
+        case Ty_any:
         case Ty_forwardPtr:
         case Ty_procPtr:
         case Ty_toLoad: